Update the output to use the correct three-dot syntax
TIL there's a semantic difference between the commit intervals a100000..b200000 and a100000...b200000. It isn't a big deal that the script outputs the wrong one, but it costs nothing to remain consistent.

If you're curious: two dot diffs compare the most recent commits on both branches, three dot diffs go back to what's in common. See https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-comparing-branches-in-pull-requests#three-dot-and-two-dot-git-diff-comparisons